Fix reserved identifiers in include guards
The include guards have been changed to the filename in uppercase
letters with all non-alphanumeric symbols replaced by underscore.
This include guard pattern matches the guard that is used for the
generated file rocsparse-export.h.

The C and C++ standards reserve all identifiers that begin with an
underscore followed by a capital letter [C99 §7.1.3]
[C++11 §17.6.4.3.2].

There's also no reason to use #pragma once and include guards together
in the same file. To quote Microsoft's documentation on #pragma once
in VS2015:

> There is no advantage to use of both the #include guard idiom and
> #pragma once in the same file. The compiler recognizes the #include
> guard idiom and implements the multiple include optimization the
> same way as the #pragma once directive if no non-comment code or
> preprocessor directive comes before or after the standard form of
> the idiom