[hipsparse] Move examples in documentation to standalone cpp files (#1318)
## Motivation

Move examples in documentation to standalone C++ files in order to allow
testing compilation of documentation examples. This is important as
otherwise when using hard-coded documentation examples, these can get
out of date with API changes and go unnoticed. Now we can compile (and
even run) our documentation examples to ensure that what is documented
is correct and working.

## Technical Details

This PR moves all of the hard-coded doxygen examples into standalone C++
programs that are then compiled as part of building hipSPARSE. These are
then included into the documentation using doxygens `\snippet`
mechanism.

## Test Plan

The examples are built as part of building hipSPARSE so as long as
hipSPARSE and the documentation can be built, then the PR works

## Test Result

N/A

## Submission Checklist

N/A