[hipsparse] Part 1: Adding yaml test files to hipSPARSE (#1326)
## Motivation

The goal is to add yaml files for testing in hipSPARSE similar to how we
handle testing in rocSPARSE. This makes the testing cleaner and more
flexible. Doing this all in one PR however would result in a very large
number of file changes and file additions. In order to keep the PR size
more manageable, I am refactoring the hipSPARSE client code to support
yaml file testing in parts:

- Part 1: Remove `std::string` from `Arguments` struct and only use
fixed sized `char` arrays
- Part 2: Add code required to process yaml file testing, generate
binary data test file, and handle the dispatching of tests.
- Part 3: Move level1 routines over to using yaml files
- Part 4: Move level2 routines over to using yaml files
- Part 5: Move level3 routines over to using yaml files
- Part 6: Move extra routines over to using yaml files
- Part 7: Move precond routines over to using yaml files
- Part 8: Move generic routines over to using yaml files
- Part 9: Move conversion routines over to using yaml files

This first PR handles Part 1, namely removing `std::string` from
`Arguments` struct.

## Technical Details

We need to remove usage of `std::string` from `Arguments` struct because
we need the `Arguments` struct to have a fixed size so that we can
properly read the binary data test file (to be named
`hipsparse_test.data`) into an `Arguments` struct object. See
https://github.com/ROCm/rocm-libraries/blob/8f896527cb01ca7dfe52b3e72d2b2fbc82a2fc31/projects/rocsparse/clients/include/rocsparse_arguments.hpp#L183
for an example of where this occurs in rocSPARSE.

## Test Plan

Regular CI tests will run. There is no expected change in test output
from the develop branch and this PR.

## Test Result

See CI tests